+++ mobile/babel.config.js
@@ -3,7 +3,8 @@ module.exports = function(api) {
   return {
     presets: ['babel-preset-expo'],
     "plugins": [
-      ["module:react-native-dotenv"]
+      ["module:react-native-dotenv"],
+      ['react-native-paper/babel']
     ]
   };
 };
