+++ mobile/package.json
@@ -21,6 +21,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.72.6",
+    "react-native-paper": "^5.11.2",
     "react-native-web": "~0.19.6",
     "typescript": "^5.1.3"
   },
